refactor(fts_column): reorganize into tokenizer/, posting/, iterator/ subdirs
