+++ src/db/CMakeLists.txt
@@ -19,7 +19,7 @@ if(NOT ANDROID AND AUTO_DETECT_ARCH)
     if(HOST_ARCH MATCHES "^(x86|x64)$")
         setup_compiler_march_for_x86(_DB_MARCH_SSE _DB_MARCH_AVX2 _DB_MARCH_AVX512 _DB_MARCH_AVX512FP16)
         set_source_files_properties(
-            ${CMAKE_CURRENT_SOURCE_DIR}/index/column/fts_column/bitpacked_simd_sse41.cc
+            ${CMAKE_CURRENT_SOURCE_DIR}/index/column/fts_column/posting/bitpacked_simd_sse41.cc
             PROPERTIES COMPILE_FLAGS "${_DB_MARCH_SSE}"
         )
     endif()
