+++ src/db/index/CMakeLists.txt
@@ -5,7 +5,7 @@ if(NOT ANDROID AND AUTO_DETECT_ARCH)
     if (HOST_ARCH MATCHES "^(x86|x64)$")
         setup_compiler_march_for_x86(INDEX_MARCH_FLAG_SSE INDEX_MARCH_FLAG_AVX2 INDEX_MARCH_FLAG_AVX512 INDEX_MARCH_FLAG_AVX512FP16)
         set_source_files_properties(
-            ${CMAKE_CURRENT_SOURCE_DIR}/column/fts_column/bitpacked_simd_sse41.cc
+            ${CMAKE_CURRENT_SOURCE_DIR}/column/fts_column/posting/bitpacked_simd_sse41.cc
             PROPERTIES
             COMPILE_FLAGS "${INDEX_MARCH_FLAG_SSE}"
         )
@@ -14,7 +14,7 @@ endif()
 
 cc_library(
     NAME zvec_index STATIC STRICT
-    SRCS *.cc segment/*.cc column/vector_column/*.cc column/inverted_column/*.cc column/fts_column/*.cc storage/*.cc storage/wal/*.cc common/*.cc
+    SRCS *.cc segment/*.cc column/vector_column/*.cc column/inverted_column/*.cc column/fts_column/*.cc column/fts_column/tokenizer/*.cc column/fts_column/posting/*.cc column/fts_column/iterator/*.cc storage/*.cc storage/wal/*.cc common/*.cc
     LIBS zvec_common
          zvec_proto
          rocksdb
