+++ src/db/index/column/fts_column/iterator/fts_conjunction_iterator.cc
