+++ src/db/index/column/fts_column/iterator/fts_disjunction_iterator.cc
