+++ src/db/index/column/fts_column/iterator/fts_phrase_iterator.cc
@@ -15,7 +15,7 @@
 #include "fts_phrase_iterator.h"
 #include <algorithm>
 #include <cstring>
-#include "fts_utils.h"
+#include "../fts_utils.h"
 
 namespace zvec::fts {
