+++ src/db/index/column/fts_column/iterator/fts_term_iterator.cc
@@ -16,7 +16,7 @@
 #include <cstring>
 #include <roaring/roaring.h>
 #include <zvec/ailego/logger/logger.h>
-#include "fts_utils.h"
+#include "../fts_utils.h"
 
 namespace zvec::fts {
