+++ src/db/index/column/fts_column/posting/bitpacked_posting_list.h
@@ -17,7 +17,7 @@
 #include <cstdint>
 #include <cstring>
 #include <string>
-#include "bm25_scorer.h"
+#include "../bm25_scorer.h"
 
 namespace zvec::fts {
