+++ src/db/index/column/fts_column/posting/bitpacked_simd_scalar.h
