+++ src/db/index/column/fts_column/tokenizer/whitespace_tokenizer.h
